# internal/api/docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: VoHive 主服务控制面 API — 自动生成 OpenAPI 文档 POC
  title: VoHive Main API
  version: "1.0"
paths:
...
  /traffic/analysis:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: TrafficAnalysis
      tags:
      - dashboard
  /upstream-proxies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ListUpstreamProxies
      tags:
      - upstream-proxies
  /upstream-proxies/{proxy_id}:
    delete:
      parameters:
      - description: proxy_id
        in: path
        name: proxy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeleteUpstreamProxy
      tags:
      - upstream-proxies
    put:
      consumes:
      - application/json
      parameters:
      - description: proxy_id
        in: path
        name: proxy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: UpdateUpstreamProxy
      tags:
      - upstream-proxies
  /upstream-proxies/{proxy_id}/lookup:
    get:
      parameters:
      - description: proxy_id
        in: path
        name: proxy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: LookupUpstreamProxy
      tags:
      - upstream-proxies
  /upstream-proxy-countries:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ListUpstreamProxyCountries
      tags:
      - upstream-proxies
  /upstream-proxy-country-rules/{country_code}:
    delete:
      parameters:
      - description: country_code
        in: path
        name: country_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeleteUpstreamProxyCountryRule
      tags:
      - upstream-proxies
    put:
      consumes:
      - application/json
      parameters:
      - description: country_code
        in: path
        name: country_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: UpsertUpstreamProxyCountryRule
      tags:
      - upstream-proxies
securityDefinitions:
  BearerAuth:
    description: '"Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"